api/swagger: remove temporary "full" example for image config
This example was added in 5e0e34fafdb4cd9904aac420dbb6dd4aa11c04c8 so that
the deprecated fields could be omitted from the example. Those fields were
removed from the swagger in 4dc961d0e922381b8cd5e05223f172f3145c7494, but
the temporary example was not removed.

This patch removes the example, in favor of the per-field examples, which
were already in place.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>