api/types/container: remove deprecated ExecOptions.Detach
This field was deprecated in 0c182d4d57e30e77fd940201205db20403fc7282,
which should be included in a 28.x release, but we don't need to carry
it in the new module.

We should also considering duplicating the `ExecOptions` type as a client
option, and renaming it to `ExecCreateRequest`, so that we can decouple
client options from the shape of the request.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>