client: remove version-gate for "--force" on "volume remove"
The `force` option on volume remove was added in [moby@6c5c34d] (docker
1.13.0-rc1, API v1.25), but did not gate the feature to API version, so
effectively introduced it to all existing API versions. After this,
[moby@e98e4a7] enabled experimental features by default, and added API
version gates, but only did so on the client side, so the daemon / API
server would continue to accept the `force` option on any API version.

Let's remove this code, given that:

- API v1.24 is the oldest API version we still handle, and only as fallback.
- This code silently discards the user's option (no warning / error)
- Every current version of the daemon handles the option, regardless
  of API version (only a 9+ year old daemon wouldn't handle it).

[moby@6c5c34d]: https://github.com/moby/moby/commit/6c5c34d50d377d1c5318a255240fb2dc9c23cf92
[moby@e98e4a7]: https://github.com/moby/moby/commit/e98e4a71110fd33852bb755a9b8b4ebc9df904db

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>