client: remove "version" header for service create, update
The version header is no longer used since [moby@a9d2091] (v20.10.0-beta1)
which was not gated by API version, as handling of the header was broken
(using the client version, instead of the API version used for the request).

Given that any current version of the daemon, regardless of API version will
ignore the header, this code was only in place to allow connecting to a
daemon older than (v20.10.0-beta1), which would be long EOL now.

[moby@a9d2091]: https://github.com/moby/moby/commit/a9d20916c3d1f5e3c3ddada79af19aee4ec64629

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>