client: ImageList: don't discard reference filter on API < 1.25
the "reference" filter was introduced in [moby@820b809] (docker 1.13.0-rc1)
to replace the "filter" query argument. That commit initially included a
version-gate anticipating the API version to be used for v17.12, but as
this was yet unknown, the version-gate was removed in [moby@0f9d22c].
A later PR re-introduced a version-gate in [moby@4a19009], reflecting the
API version in which the deprecation was (finally) completed.

For the client, [moby@c6e3145] added a fallback was added for older daemons
(docker 1.12.0 and older, using API < v1.25) that did not support the new
filter.

Looking at the above, any version of docker 1.13.0 or above handles the
"reference" filter, but (depending on the docker version) may also handle
the old filter on API < 1.28 or API < 1.41. Removing this option will only
impact daemon versions older than 1.13.0, which are long obsolete.

Given that current clients forcibly remove the "reference" filter and replace
it with the old "filter" when using API v1.24, we keep support on the daemon
side, but update the version to v1.24, and only if no reference filter is
set.

[moby@820b809]: https://github.com/moby/moby/commit/820b809e70df8b9c7af00256182c48d935972a5c
[moby@c6e3145]: https://github.com/moby/moby/commit/c6e31454ba2f053bc6831651663cf538d142afaa
[moby@0f9d22c]: https://github.com/moby/moby/commit/0f9d22cd66353b3d14dd4a08084f88778fb69480
[moby@4a19009]: https://github.com/moby/moby/commit/4a1900915a2eb80f7eaf3638ed702b5d80bfa80f

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>