client: Change ImageRemove to accept handle.ImageHandle parameter
The ImageRemove method signature has been updated to accept a
handle.ImageHandle interface instead of a string for the image
parameter.

This change introduces a new abstraction layer for image
references that allows for more flexible image resolution strategies.

Currently only a naive image resolution is available that directly
corresponds to the previous behavior of just passing a string image
reference.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>