image inspect: fix legacy fields for API < v1.52 response
This was a mistake I made in eafca64a6bb6fda72cfe07c294780e65173de1ce,
which was extracted from a set of changes that I had in progress, but
I forgot that patch was unfinished (whoops!); before that commit, the
legacy "Config" fields were handled through the `inspectCompatResponse`,
which applied the legacy fields to the `Config` struct within the image
inspect response.

When changing the implementation to use the `compat` package, those
fields were applied at the top-level of the response, instead of the
`Config`; additional changes were needed for the `compat` package to
support pathing nested structs, and to prevent "extra" fields from
overwriting fields that already existed in the response; these changes
were implemented in e204ba1dcaba62beaadd1d95574f8784d9c4dd7c.

This patch:

- Removes the old `inspectCompatResponse` implementation, which was
  no longer used.
- Updates the router to patch the `Config` fields, using the fixes
  and enhancements that were implemented in e204ba1dcaba62beaadd1d95574f8784d9c4dd7c.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>