api/types/image: InspectResponse: deprecate Author
The `Author` field can be set through the `MAINTAINER` instruction in
Dockerfiles, and through the `--author` option on `docker commit`, but
is optional, and won't be set in most situations.

This patch deprecates the fields in the `InspectResponse` go struct. The
`Author` field can still be set but is deprecated in the Dockerfile syntax,
so should not be depended on. This field was implicitly deprecated when
the `MAINTAINER` command was deprecated in [moby@efb9e38] / [moby@6694974]
(docker 1.13, API v1.25), however the related API fields were kept so
that information of legacy images would not be discarded.

The API continues to return the field if set, allowing the client to
print the fields for informational purposes when printing the raw response,
but these fields should be considered "transitional", and not be depended
on; deprecating the fields helps raise awareness.

[moby@efb9e38]: https://github.com/moby/moby/commit/efb9e38ebab4b60a73e3295e3c468813182fd3ea
[moby@6694974]: https://github.com/moby/moby/commit/6694974c9e3532cdaf5333b0a8b19f5ca64c70f0

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>